+++ blob/service/client/client.go
@@ -43,7 +43,7 @@ func New(provider Provider) (*Client, error) {
 
 // FUTURE: Return ErrorResourceNotFoundWithID(userID) if userID does not exist at all
 
-func (c *Client) List(ctx context.Context, userID string, filter *blob.Filter, pagination *page.Pagination) (blob.Blobs, error) {
+func (c *Client) List(ctx context.Context, userID string, filter *blob.Filter, pagination *page.Pagination) (blob.BlobArray, error) {
 	if err := c.AuthClient().EnsureAuthorizedService(ctx); err != nil {
 		return nil, err
 	}