+++ blob/store/structured/structured.go
@@ -19,7 +19,7 @@ type Store interface {
 type Session interface {
 	io.Closer
 
-	List(ctx context.Context, userID string, filter *blob.Filter, pagination *page.Pagination) (blob.Blobs, error)
+	List(ctx context.Context, userID string, filter *blob.Filter, pagination *page.Pagination) (blob.BlobArray, error)
 	Create(ctx context.Context, userID string, create *Create) (*blob.Blob, error)
 	Get(ctx context.Context, id string) (*blob.Blob, error)
 	Update(ctx context.Context, id string, condition *request.Condition, update *Update) (*blob.Blob, error)