+++ data/source/service/client/client.go
@@ -33,7 +33,7 @@ func New(provider Provider) (*Client, error) {
 
 // FUTURE: Return ErrorResourceNotFoundWithID(userID) if userID does not exist at all
 
-func (c *Client) List(ctx context.Context, userID string, filter *dataSource.Filter, pagination *page.Pagination) (dataSource.Sources, error) {
+func (c *Client) List(ctx context.Context, userID string, filter *dataSource.Filter, pagination *page.Pagination) (dataSource.SourceArray, error) {
 	if _, err := c.AuthClient().EnsureAuthorizedUser(ctx, userID, permission.Owner); err != nil {
 		return nil, err
 	}